Rename to minimnize transactions
