+++ src/payout/transaction.py
@@ -1,0 +1,3 @@
+from collections import namedtuple
+
+Transaction = namedtuple("Transaction", ["sender_id", "receiver_id", "cents"])
